enlarge premium rate precision

+++ contracts/interfaces/IOptionVaultManager.sol
@@ -17,7 +17,7 @@ interface IOptionVaultManager {
     function buyOptions(uint8[] memory _vaultIds) payable external; 
     function whitelistTraders() external view returns(address[] memory);
 
-    event OptionBought(uint8 indexed _vaultId, uint16 indexed _currentRound, address indexed _buyerAddress, uint256 _amount, uint128 _strike, uint16 _premiumRate);
+    event OptionBought(uint8 indexed _vaultId, uint16 indexed _currentRound, address indexed _buyerAddress, uint256 _amount, uint128 _strike, uint104 _premiumRate);
     
     event OptionExpired(uint8 indexed _vaultId, uint16 indexed _currentRound, uint128 _expiryLevel, uint256 _optionHolderValue);
